refactor: add eslint rule for test files, remove all eslint warnings, rename apiservice, build url using search params
