+++ src/components/Pagination/__tests__/index.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen, fireEvent } from '@testing-library/react';
 import { vi, describe, it, beforeEach, expect } from 'vitest';
-import { StubProvider } from '../../../router';
+import { StubProvider } from '../../../router/utils';
 import Pagination from '..';
 
 describe('Pagination Component', () => {
@@ -28,6 +28,7 @@ describe('Pagination Component', () => {
     expect(nextPaginationButton).toBeInTheDocument();
     expect(prevPaginationButton).toBeInTheDocument();
   });
+
   it('render first ...', async () => {
     render(
       <StubProvider
@@ -87,6 +88,7 @@ describe('Pagination Component', () => {
 
     expect(manyDotesComponent).toBeInTheDocument();
   });
+
   it('click previous button', async () => {
     let page = 8;
 
@@ -114,6 +116,7 @@ describe('Pagination Component', () => {
 
     expect(page).toBe(7);
   });
+
   it('click next button', async () => {
     let page = 8;
     render(
@@ -140,6 +143,7 @@ describe('Pagination Component', () => {
 
     expect(page).toBe(9);
   });
+
   it('click previous mobile button', async () => {
     let page = 8;
 
@@ -167,6 +171,7 @@ describe('Pagination Component', () => {
 
     expect(page).toBe(7);
   });
+
   it('click next mobile button', async () => {
     let page = 8;
     render(
@@ -193,6 +198,7 @@ describe('Pagination Component', () => {
 
     expect(page).toBe(9);
   });
+
   it('click first page button', async () => {
     let page = 5;
     render(
@@ -219,6 +225,7 @@ describe('Pagination Component', () => {
 
     expect(page).toBe(1);
   });
+
   it('click first page button', async () => {
     let page = 5;
     render(
@@ -245,6 +252,7 @@ describe('Pagination Component', () => {
 
     expect(page).toBe(1);
   });
+
   it('click last page button', async () => {
     let page = 5;
 
@@ -276,6 +284,7 @@ describe('Pagination Component', () => {
 
     expect(page).toBe(totalPages);
   });
+
   it('click digit button (4)', async () => {
     let page = 5;
