+++ src/pages/About/__tests__/index.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import { describe, it, expect } from 'vitest';
-import { Stub } from '../../../router';
+import { Stub } from '../../../router/utils';
 
 describe('About page', () => {
   it('displays text correctly', () => {
