+++ src/pages/NotFound/__tests__/index.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import { describe, it, expect } from 'vitest';
-import { Stub } from '../../../router';
+import { Stub } from '../../../router/utils';
 
 describe('Not found page', () => {
   it('displays 404 text if page is not exists', () => {
